pyproject.toml
[project]
name = "agriassist"
version = "0.1.0"
description = "AI-powered farming assistant providing multimodal advice through text, image, and audio inputs"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "AgriAssist Team", email = "team@agriassist.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.5",
    "groq>=0.31.1",
    "gtts>=2.5.4",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "aiofiles>=24.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "I", "N", "UP", "YTT", "BLE", "B", "A", "C4", "PIE", "T20", "SIM", "ARG", "PTH", "PLR", "PLW", "TRY", "NPY", "RUF"]
ignore = ["E501", "PLR0913", "TRY003"]

[tool.black]
line-length = 88
target-version = ['py312']
